+++ testdata/de/dm/intellij/liferay/language/freemarker/LiferayFtlVariableProviderTest/de/dm/MyEnum.java
@@ -1,0 +1,8 @@
+package de.dm;
+
+public enum MyEnum {
+
+    FOO,
+    BAR
+
+}