+++ testdata/de/dm/intellij/liferay/language/freemarker/LiferayFtlVariableProviderTest/de/dm/MyUtil.java
@@ -1,0 +1,8 @@
+package de.dm;
+
+public class MyUtil {
+
+    public static void sayHello() {
+
+    }
+}